# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kroków — API draft",
    "version": "1.0.0-draft.2",
    "description": "Draft for A+B review. Schemas do not imply deployed endpoints. Cross-field and catalog reference constraints require the exported Zod validators."
  },
  "paths": {
...
    "/v1/coverage": {
      "get": {
        "operationId": "coverage",
        "responses": {
          "200": {
            "description": "Text and spatial coverage",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoverageResponse"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "DATA_VERSION_CHANGED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "SAME_ENDPOINT, OUTSIDE_COVERAGE, UNRESOLVED_ENDPOINT, NO_PATH, NO_MATCHING_ROUTE, FEATURE_NOT_ENABLED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds before retry; matches error.details.retryAfterSec",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "SOURCE_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "504": {
            "description": "SEARCH_LIMIT_REACHED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/places": {
      "get": {
        "operationId": "places",
        "parameters": [
          {
            "name": "cityId",
            "in": "query",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "minLength": 1,
              "maxLength": 160
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "anyOf": [
                {
                  "type": "string",
                  "const": ""
                },
                {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 100
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Up to ten catalog places",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlacesResponse"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "DATA_VERSION_CHANGED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "SAME_ENDPOINT, OUTSIDE_COVERAGE, UNRESOLVED_ENDPOINT, NO_PATH, NO_MATCHING_ROUTE, FEATURE_NOT_ENABLED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds before retry; matches error.details.retryAfterSec",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "SOURCE_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "504": {
            "description": "SEARCH_LIMIT_REACHED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/place-search": {
      "get": {
        "operationId": "placeSearch",
        "parameters": [
          {
            "name": "cityId",
            "in": "query",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "minLength": 1,
              "maxLength": 160
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "anyOf": [
                {
                  "type": "string",
                  "const": ""
                },
                {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 100
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "City-wide search; routing availability is explicit for each result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceSearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "DATA_VERSION_CHANGED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "SAME_ENDPOINT, OUTSIDE_COVERAGE, UNRESOLVED_ENDPOINT, NO_PATH, NO_MATCHING_ROUTE, FEATURE_NOT_ENABLED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds before retry; matches error.details.retryAfterSec",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "SOURCE_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "504": {
            "description": "SEARCH_LIMIT_REACHED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/routes": {
      "post": {
        "operationId": "routes",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "One to three routes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponse"
                }
              }
            },
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "DATA_VERSION_CHANGED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "SAME_ENDPOINT, OUTSIDE_COVERAGE, UNRESOLVED_ENDPOINT, NO_PATH, NO_MATCHING_ROUTE, FEATURE_NOT_ENABLED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds before retry; matches error.details.retryAfterSec",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "SOURCE_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "504": {
            "description": "SEARCH_LIMIT_REACHED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sources": {
      "get": {
        "operationId": "sources",
        "responses": {
          "200": {
            "description": "Public sources",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcesResponse"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "DATA_VERSION_CHANGED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "SAME_ENDPOINT, OUTSIDE_COVERAGE, UNRESOLVED_ENDPOINT, NO_PATH, NO_MATCHING_ROUTE, FEATURE_NOT_ENABLED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds before retry; matches error.details.retryAfterSec",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "SOURCE_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "504": {
            "description": "SEARCH_LIMIT_REACHED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/features/{id}/evidence": {
      "get": {
        "operationId": "featureEvidence",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "minLength": 1,
              "maxLength": 160
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Published redacted observations only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureEvidenceResponse"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "DATA_VERSION_CHANGED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "SAME_ENDPOINT, OUTSIDE_COVERAGE, UNRESOLVED_ENDPOINT, NO_PATH, NO_MATCHING_ROUTE, FEATURE_NOT_ENABLED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds before retry; matches error.details.retryAfterSec",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "SOURCE_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "504": {
            "description": "SEARCH_LIMIT_REACHED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CoverageResponse": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "1.0.0-draft.2"
          },
          "requestId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "mode": {
            "type": "string",
            "enum": [
              "synthetic",
              "pilot"
            ]
          },
          "navigationEligibility": {
            "type": "string",
            "enum": [
              "preview_only",
              "foreground_experimental"
            ]
          },
          "cityId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "capabilities": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "catalog_routes",
                "coordinate_endpoints",
                "foreground_navigation"
              ]
            }
          },
          "polygon": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "const": "Polygon"
                  },
                  "coordinates": {
                    "minItems": 1,
                    "type": "array",
                    "items": {
                      "minItems": 4,
                      "type": "array",
                      "items": {
                        "type": "array",
                        "prefixItems": [
                          {
                            "type": "number",
                            "minimum": -180,
                            "maximum": 180
                          },
                          {
                            "type": "number",
                            "minimum": -90,
                            "maximum": 90
                          }
                        ],
                        "items": false,
                        "minItems": 2,
                        "maxItems": 2
                      }
                    }
                  }
                },
                "required": [
                  "type",
                  "coordinates"
                ],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ]
          },
          "policy": {
            "type": "object",
            "properties": {
              "policyVersion": {
                "type": "string",
                "minLength": 1,
                "maxLength": 160
              },
              "fieldVerificationMaxAgeDays": {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "policyVersion",
              "fieldVerificationMaxAgeDays"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "schemaVersion",
          "requestId",
          "mode",
          "navigationEligibility",
          "cityId",
          "name",
          "description",
          "capabilities",
          "polygon",
          "policy"
        ],
        "additionalProperties": false
      },
...
      "PlacesResponse": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "1.0.0-draft.2"
          },
          "requestId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "mode": {
            "type": "string",
            "enum": [
              "synthetic",
              "pilot"
            ]
          },
          "navigationEligibility": {
            "type": "string",
            "enum": [
              "preview_only",
              "foreground_experimental"
            ]
          },
          "cityId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "places": {
            "maxItems": 10,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 160
                },
                "cityId": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 160
                },
                "name": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 500
                },
                "description": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 500
                },
                "coordinate": {
                  "anyOf": [
                    {
                      "type": "array",
                      "prefixItems": [
                        {
                          "type": "number",
                          "minimum": -180,
                          "maximum": 180
                        },
                        {
                          "type": "number",
                          "minimum": -90,
                          "maximum": 90
                        }
                      ],
                      "items": false,
                      "minItems": 2,
                      "maxItems": 2
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              },
              "required": [
                "id",
                "cityId",
                "name",
                "description",
                "coordinate"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "schemaVersion",
          "requestId",
          "mode",
          "navigationEligibility",
          "cityId",
          "places"
        ],
        "additionalProperties": false
      },
...
    }
  }
}